remove lib
