+++ src/data_collection/processes/linux.rs
@@ -18,7 +18,7 @@ use crate::{
     utils::error::{self, BottomError},
 };
 
-/// Maximum character length of a /proc/<PID>/stat process name.
+/// Maximum character length of a `/proc/<PID>/stat`` process name.
 /// If it's equal or greater, then we instead refer to the command for the name.
 const MAX_STAT_NAME_LEN: usize = 15;
